Updates based on PR comments

+++ test/include/test.h
@@ -24,8 +24,8 @@
     uint64_t start_time_ms = sys_timestamp_ms();                               \
     bool success = ENTRYPOINT();                                               \
     uint64_t elapsed_time_ms = sys_timestamp_ms() - start_time_ms;             \
-    sys_printf("TEST %s (%s) in %" PRIu64 " ms\n", success ? "PASS" : "FAIL",  \
-               sys_env_name(), elapsed_time_ms);                               \
+    sys_printf("TEST %s (%s) in %llu ms\n", success ? "PASS" : "FAIL",         \
+               sys_env_name(), (unsigned long long)elapsed_time_ms);           \
     hw_exit();                                                                 \
     sys_exit();                                                                \
     return success ? 0 : 1;                                                    \
